This is the python script containing the classes that i use to perform the data scraping.
The code implements diverse functions for scraping data from YouTube and subsequently storing the acquired information in a relational database. Specifically, a relational database is established to serve as the repository for the scraped data. It is imperative to replace the placeholder string "YOU API KEY" on line 9 with the actual YouTube API key for proper functionality.